+++ .eslintrc.js
@@ -110,7 +110,8 @@ module.exports = {
                 '@typescript-eslint/no-unnecessary-qualifier': 'error',
                 '@typescript-eslint/no-unused-vars': [
                     'error', {
-                        argsIgnorePattern: '^_'
+                        argsIgnorePattern: '^_',
+                        varsIgnorePattern: '^_'
                     }
                 ],
                 '@typescript-eslint/no-useless-constructor': 'error',