+++ src/tsp-command-types.ts
@@ -173,3 +173,25 @@ export enum ScriptElementKind {
     /** Jsdoc @link: in `{@link C link text}`, the link text "link text" */
     linkText = 'link text'
 }
+
+export class KindModifiers {
+    public static readonly optional = 'optional';
+    public static readonly deprecated = 'deprecated';
+    public static readonly color = 'color';
+
+    public static readonly dtsFile = '.d.ts';
+    public static readonly tsFile = '.ts';
+    public static readonly tsxFile = '.tsx';
+    public static readonly jsFile = '.js';
+    public static readonly jsxFile = '.jsx';
+    public static readonly jsonFile = '.json';
+
+    public static readonly fileExtensionKindModifiers = [
+        KindModifiers.dtsFile,
+        KindModifiers.tsFile,
+        KindModifiers.tsxFile,
+        KindModifiers.jsFile,
+        KindModifiers.jsxFile,
+        KindModifiers.jsonFile
+    ];
+}